requirements.txt
--index-url https://pypi.org/simple
bs4
clipboard
colorama
darglint
EbookLib
genanki
google-auth
google-auth-stubs
gprof2dot
gspread
json5
markdown
matplotlib
mypy
oauth2client
openpyxl
pandas
pandas-stubs
pathlib
Pillow
pip-review
pre-commit
setuptools
types-beautifulsoup4
types-colorama
types-requests
types-setuptools
typing
urllib3
